# docs/openapi.yaml
openapi: 3.0.3
info:
  title: jjobb_v2 Graduate Network API
  version: 2.0.0-sprint4
  description: Sprint 4 recommendations, field trips, networking school scope (+ Sprint 2–3). v1 `/api` keys are preserved.
servers:
  - url: http://localhost:5000
    description: Local / Docker Compose backend
...
paths:
...
  /api/files/{id}:
    get:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: Download stored file (school-scoped)
      responses:
        '200': { description: binary }
        '403': { description: FORBIDDEN }
  /api/jobs/{id}/apply:
    post:
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      summary: Apply; accepts resume_id (REQ-JOB-005)
      responses:
        '201': { description: application }
  /api/jobs/applications/{id}/status:
    patch:
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      summary: Application workflow status (REQ-JOB-003/004)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, reviewed, interviewed, accepted, rejected]
      responses:
        '200': { description: updated }
        '400': { description: VALIDATION }
        '403': { description: FORBIDDEN }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT